refactor(ui): migrate CardLink component to shadcn
