+++ apps/next/app/roles/[slug]/(public)/page.tsx
@@ -3,9 +3,9 @@ import { InformationCircleIcon } from "@heroicons/react/24/outline";
 import Image from "next/image";
 import Link from "next/link";
 import { notFound, useParams } from "next/navigation";
-import CardLink from "@/components/CardLink";
 import SimpleLayout from "@/components/layouts/Simple";
 import Placeholder from "@/components/Placeholder";
+import { CardLink } from "@/components/ui/card-link";
 import logo from "@/images/flexile-logo.svg";
 import { trpc } from "@/trpc/client";
 import { toSlug } from "@/utils";
