+++ apps/next/app/updates/company/page.tsx
@@ -3,7 +3,6 @@ import { CheckCircleIcon, TrashIcon } from "@heroicons/react/24/outline";
 import Link from "next/link";
 import { useRouter } from "next/navigation";
 import React, { useMemo, useState } from "react";
-import CardLink from "@/components/CardLink";
 import MainLayout from "@/components/layouts/Main";
 import Modal from "@/components/Modal";
 import MutationButton from "@/components/MutationButton";
@@ -12,6 +11,7 @@ import Placeholder from "@/components/Placeholder";
 import Status from "@/components/Status";
 import Table, { createColumnHelper, useTable } from "@/components/Table";
 import { Button } from "@/components/ui/button";
+import { CardLink } from "@/components/ui/card-link";
 import { useCurrentCompany, useCurrentUser } from "@/global";
 import { trpc } from "@/trpc/client";
 import { formatDate } from "@/utils/time";
